Change all primitive to linked types

+++ CommonJPA/src/main/java/ru/yanes/autoprom/projections/GenerationShortView.java
@@ -6,7 +6,7 @@ import ru.yanes.autoprom.entity.GenerationPhoto;
 import java.util.Set;
 
 public interface GenerationShortView {
-	int getId();
+	Short getId();
 	@Value("#{target.model.brand.shortName + ' ' + target.model.fullName + ' ' + target.fullName}")
 	String getFullName();
 	@Value("#{target.photos}")
